Resolved payment templates
