+++ apps/web/src/templates/payment-failure-template.tsx
@@ -1,25 +1,30 @@
 import { useRouter } from 'next/router'
 import useTranslation from 'next-translate/useTranslation'
+import { useCallback } from 'react'
 
 import EmailVerification from '@/components/profile/email-verification'
-import Success from '@/components/purchase-form/shared/success'
+import Failure from '@/components/purchase-form/shared/failure'
 import { useAuth } from '@/contexts/auth-context'
 import { urls } from '@/utils/urls'
 
-export default function CheckoutSuccessTemplate() {
+export default function PaymentFailureTemplate() {
   const { t } = useTranslation()
   const { user } = useAuth()
   const { push } = useRouter()
+
+  const handleRetry = useCallback(() => {
+    push(urls.releases)
+  }, [push])
 
   if (!user?.emailVerified) {
     return <EmailVerification inline />
   }
   return (
-    <Success
-      buttonText={t('common:actions.View My Collection')}
-      handleClick={() => push(urls.myCollection)}
-      headingClassName="mb-16"
-      headingText={t('common:statuses.Success!')}
+    <Failure
+      buttonText={t('common:actions.Try Again')}
+      error={t('forms:errors.failedPayment')}
+      handleClick={handleRetry}
+      headingText={t('release:failedToClaim')}
     />
   )
 }
